refactor: update utility imports and improve component structure
