+++ components/Hero.tsx
@@ -8,40 +8,45 @@ const Hero = () => {
     return (
         <div className={'pb-20 pt-36'}>
             <div>
-                <Spotlight
-                    className="-top-40 -left-10 md:-left-32 md:-top-20 h-screen"
-                    fill="white"
-                />
-                <Spotlight
-                    className="h-[80vh] w-[50vw] top-10 left-full"
-                    fill="purple"
-                />
-                <Spotlight className="left-80 top-28 h-[80vh] w-[50vw]" fill="blue" />
+                {/*Spotlight Effect*/}
+                <Spotlight className="-top-40 -left-10 md:-left-32 md:-top-20 h-screen" fill="white" />
+                <Spotlight className="top-10 left-full h-[80vh] w-[50vw]" fill="purple" />
+                <Spotlight className="top-28 left-80 h-[80vh] w-[50vw]" fill="blue" />
 
-                <div className="h-screen w-full dark:bg-black-100 bg-white dark:bg-grid-white/[0.03] bg-grid-black-100/[0.2] absolute top-0 left-0 flex items-center justify-center">
+                {/*Grid Background*/}
+                <div className="h-screen w-full dark:bg-black-100 bg-white dark:bg-grid-white/[0.03] bg-grid-black-100/[0.2]
+                absolute top-0 left-0 flex items-center justify-center">
 
-                    <div className="absolute pointer-events-none inset-0 flex items-center justify-center dark:bg-black-100 bg-white [mask-image:radial-gradient(ellipse_at_center,transparent_20%,black)]"/>
+                    <div className="absolute pointer-events-none inset-0 flex items-center justify-center dark:bg-black-100 bg-white
+                    [mask-image:radial-gradient(ellipse_at_center,transparent_20%,black)]"/>
 
                 </div>
 
+                {/*Hero Content*/}
                 <div className={'flex justify-center relative my-20 z-10'}>
                     <div className={'max-w-[89vw] md:max-w-2xl lg:max-w-[60vw] flex flex-col items-center justify-center'}>
-                        <h2 className={'uppercase tracking-widest text-xs text-center text-blue-100 max-w-80'}> Dynamic Web Magic with Next.js
+
+                        {/*Subheading*/}
+                        <h2 className={'uppercase tracking-widest text-xs text-center text-blue-100 max-w-80'}>
+                            Dynamic Web Magic with Next.js
                         </h2>
+
+                        {/*Main Heading*/}
                         <TextGenerateEffect
                             words="Transforming Concepts into Seamless User Experiences"
                             className="text-center text-[40px] md:text-5xl lg:text-6xl"
                         />
-                        <p className={'text-center md:text-lg lg:text-2xl'}>
+
+                        {/*Subheading*/}
+                        <p className={'text-center md:tracking-wider mb-4 text-sm md:text-lg lg:text-2xl'}>
                             Hi, I&apos;m Gary, a Next.js developer based in Miami, FL.
                         </p>
+
+                        {/*Button*/}
                         <a href={'#about'}>
-                            <MagicButton
-                                title={'Show My Work'}
-                                icon={<FaLocationArrow />}
-                                position={'right'}
-                            />
+                            <MagicButton title={'Show My Work'} icon={<FaLocationArrow />} position={'right'} />
                         </a>
+
                     </div>
                 </div>
