+++ components/ui/FloatingNav.tsx
@@ -6,7 +6,7 @@ import {
     useScroll,
     useMotionValueEvent,
 } from "framer-motion";
-import { cn } from '@/utils/cn';
+import { cn } from '@/lib/utils';
 import Link from "next/link";
 
 export const FloatingNav = ({
